Add tests for two methods that students should write
- SockDrawer#match_for
- SockDrawer#release_pair

These two methods will require students to work with a collection of
objects.  A sock drawer has a collection of socks.  In one method, the
sock drawer has to find a match for a sock that it's given.  In the
other, the sock drawer has to produce a pair of socks found within
its collection.