Revert "change SockMatch from class to module"
This reverts commit 6dbcef2e665559b4d5a55f5123d5455e88360494.

Based on delivery in NYC this change set this
question backward.

1.  Students saw the magic word `module` and took
that to mean that the expectation was tht the
thing be included into `SockDrawer`.
1.  To those familiar with dependency injection
via POODR (which is held as a religious text in
phase 1), the will not expect to see a constant
injected, but rather an instance as was the case
in the previous release.
1. The `self` extension of a `module` is an
unusual invocation of some `module` / `class`
metaprogramming magic whose purpose is confusing
for beginners