README: Explicitly document mutation expectation in methods
Students have expressed confusion as to whether /
why `match_for` ought be destructive. They contend
that it need not _necessarily_ be a destructive
operation. Explicitly declare the destructivity in
the README.

Similarly, it's not the case that
`supply_a_random_pair` means "identify **AND**
remove" it. Set the mutation expectation
explicitly.