Remove "random" word that's confusing some students
Per #118, the word _random_ seems to push some
students to over-invest in randomization routines.

Based on usage guidance:

> [to] say that a quantity is chosen randomly, I'm
> implicitly saying that there's a probability
> measure over that set

We're not doing anything with a statistical
measure spread being calculated with respect to
foot garments, so I believe the correct word is
"arbitrary" (versus "random") and instead of a
"random" pair merely get "one" pair.