wrap method arguments in parens
The assessment is not the right time for students to be introduced to
the idea that ruby allows arguments to be passed without parens.
Further, there are a number of very high profile Ruby style guides that
err on the side of avoiding this syntax altogether, because it can be
ambiguous and lead to strange errors.

This change adds back in the optional parens for method calls on the
domain classes. I speficially left them out for rspec matcher syntax eg.
`expect(something).to include a_thing`.